speed up asserting whether IET connections actually exist
this was still a bottleneck in deformed surfaces where sectors tend to
get very thin very quickly.

+++ libflatsurf/src/impl/assert_connection.hpp
@@ -32,7 +32,7 @@ class AssertConnection {
   bool operator()(const SaddleConnection<FlatTriangulation<T>>&);
 
  private:
-  long long budget = 1024;
+  int budget = 1024;
   long long cost = 1;
 };
